Adds og hook methods to track user membership on projects.
Also makes several infrastructural improvements to the project.  The silly filepath parameter has been removed from the post_request method and replaced with a check for a filestring at data['file'].  Adds OG as a dependency of tapas_transform, which is required for invoking og api hooks.  Implements a put_request method to be used with all TapasHydraConnector update requests and factors out a general purpose http_request method to be used to support the various verbs (get/put/post/delete) that the Rails repository knows what to do with.