Test 'derive MODS' using test suite
The new test suite makes use of XQuery 3.0 annotations, the XQSuite
framework, and the EXPath HTTP Client Module. For each test defined,
the test runner will fire off an HTTP request to the correct
tapas-xq URL, and parse the response for expected values. The test
runner must be executed with 'dba' permissions.

Only 'derive MODS' has been implemented so far, and no attempt was
made to test logging in as a user without valid access permissions.