Test the program type specified by the view package config file
If the program type doesn't match those handled by
derive-reader.xq, a HTTP 501 error is returned, because that type
is unsupported. A 501 will also be returned if there is no
configuration file for the given view package. The latter case is
possible now, because the list of possible view packages is hand-
maintained rather than generated or compiled. When
$txq:valid-reader-types is made dynamic, the existence of a
configuration file will be guaranteed.

Currently, only XSLT programs are supported. XProc and XQuery are
allowed by the config file schema but aren't yet implemented.