Add function to get the packages that can be updated
The function dpkg:get-updatable() is a report-maker; it is not
intended to enable special handling when, say, all packages must
be updated (a zip file would be faster to 'install'). The function
lists all packages that are either unavailable or out-of-date.

This functionality depends on (and tests for) an XML registry of
view packages available in eXist, but there is not yet code for
eXist to build or update that registry.