Create registry and insert entries on update
Before updating a package using the latest commit on a branch,
TAPAS-xq attempts to find a commit matching the git_timestamp
field from Rails.

View package functions previously placed in general-functions.xql
are now in view-pkgs.xql, preventing infinite recursion in module
imports.

An attempt is made to update the view package registry if it had
not existed before installation.