Initialize packages by downloading a ZIP archive
TAPAS-xq now makes fewer calls to the GitHub API. Instead of using
the Contents API to get at files, TAPAS-xq uses the Archive API to
get all files at once, or the Compare Commits API to get only the
files changed between Rails' version of a package and eXist's.

Updating a package that ISN'T in a tapas-view-packages submodule
currently requires updating the entire tapas-view-packages repo at
the commit matching Rails' timestamp for the given package. While
that works for the TAPAS environment now, it could be better
implemented by getting the tapas-view-packages ZIP and then
extracting only the files matching the path to the given package.