Log in as TAPAS user before initializing view packages
After TAPAS-xq is installed, the post-install script will try to
log in as the TAPAS user, using the default password. Only if the
login succeeds will the script initialize view packages. This adds
a manual task to setting up the auto-ingest feature on the TAPAS
servers, but it also prevents the contents of the tapas-view-pkgs
directory from being owned by the admin user (and only writeable by
that user).

A private function has been added to the view packages library,
which tests if the logged-in user is the TAPAS user. However, a
bug in eXist v2.2 means that this function can't be used to ensure
that updating/initializing only occurs when the TAPAS user has
control of the process.