Create means of deciding which Rails API to use (test or prod)
An administrator can add the attribute @rails-api to explicitly
set the URL that TAPAS-xq will call to find view packages. If the
attribute is present, that URL will be used. Otherwise, TAPAS-xq
will use the Rails API on the production machine.

Also, when unpacking a ZIP archive, submodules will be identified
from the entries within the ZIP, not from the contents of the
database. This allows the complete regeneration of package(s) by
manually removing <package_ref>s from the registry without also
having to delete the local package directory.