Define an environment configuration file for eXist
When TAPAS-xq is installed, it looks for 'environment.xml' at '/db'.
If the file doesn't exist, it stores a copy of
'environment.xml.default' at the above path.

This configuration file defines the base URL eXist needs to send
requests to Rails; the default branch to use for GitHub repository
queries; and what kind of environment eXist is running in
('production', 'development-local', or 'development-staging'). In
the default file, 'development-local' is assumed.