Use custom element to pass around errors.
Also, test `tap:get-file-content()` against various inputs: a node,
a string, and an xs:base64Binary for both UTF-8 and UTF-16 encoded,
well-formed XML.

Unfortunately, the test fails when a UTF-16 document is loaded as a
string (much less if it is then encoded as binary). The same
document can be loaded as a node via `doc()`, however.