Remove `dpkg:fetch-remote-history#1`, clean up code.
Since we're using the `git` command, I'm not comfortable updating
submodules independently of the main View Package repository via
BaseX. TAPAS-xq should rely on the GitHub repository branches for
canonical information. It should be able to switch branches and
possibly check out commits, but it should not be able to make
changes that would cause it to go out of sync with the remote repo.