Limit read access when resources are marked as private.
If the user requesting a TEI, MODS, or TFE file is not the "tapas"
user, the TFE is checked to determine if the resource is publicly
available. If not, an error code is returned.

This needs more work to ensure that admin accounts have access as
well. Also more detail is needed in the API documentation.